+++ include/tools/otherTools.h
@@ -24,6 +24,7 @@
 #include <vector>
 #include <map>
 #include <zim/zim.h>
+#include <mustache.hpp>
 
 namespace pugi {
   class xml_node;
@@ -47,6 +48,9 @@ namespace kiwix
   MimeCounterType parseMimetypeCounter(const std::string& counterData);
 
   std::string gen_date_str();
+  std::string gen_uuid(const std::string& s);
+
+  std::string render_template(const std::string& template_str, kainjow::mustache::data data);
 }
 
 #endif
