+++ src/tools/otherTools.cpp
@@ -33,6 +33,8 @@
 #include <sstream>
 #include <pugixml.hpp>
 
+#include <zim/uuid.h>
+
 
 static std::map<std::string, std::string> codeisomapping {
 { "aa", "aar" },
@@ -357,4 +359,20 @@ std::string kiwix::gen_date_str()
      << std::setw(2) << std::setfill('0') << tm->tm_min << ":"
      << std::setw(2) << std::setfill('0') << tm->tm_sec << "Z";
   return is.str();
+}
+
+std::string kiwix::gen_uuid(const std::string& s)
+{
+  return kiwix::to_string(zim::Uuid::generate(s));
+}
+
+std::string kiwix::render_template(const std::string& template_str, kainjow::mustache::data data)
+{
+  kainjow::mustache::mustache tmpl(template_str);
+  kainjow::mustache::data urlencode{kainjow::mustache::lambda2{
+                               [](const std::string& str,const kainjow::mustache::renderer& r) { return urlEncode(r(str), true); }}};
+  data.set("urlencoded", urlencode);
+  std::stringstream ss;
+  tmpl.render(data, [&ss](const std::string& str) { ss << str; });
+  return ss.str();
 }
